# QueryManager: make post sorting faster and call it less often