@nylen 2017-11-17
The code looks solid and the posts list seems to be working correctly.  Subjectively, this PR cuts the delay when loading a page of posts roughly in half (4s after the API request completes down to 2s).

Overall, test coverage for `QueryManager` should be pretty good, so I also wanted to use this to try to get a feel for whether these changes might break anything.

First, I commented out the `this.constructor.sort` line and ran all tests under `client/lib/query-manager/`.  6 of them failed.  This gives me confidence that the sorting is covered by tests and that it's still working as intended.

Second, I wondered about potential differences between `+moment( post.date )` and `+new Date( post.date )`.  The only thing that came to mind was timezone handling:  the API can return posts in multiple timezones (from different sites in All My Sites mode, or when a site's timezone changes).

```js
$ node
> const moment = require( 'moment' );
undefined
> [ +moment( '2017-11-17T04:05:33-08:00' ), +moment( '2017-09-30T17:59:39+00:00' ) ]
[ 1510920333000, 1506794379000 ]
> [ +new Date( '2017-11-17T04:05:33-08:00' ), +new Date( '2017-09-30T17:59:39+00:00' ) ]
[ 1510920333000, 1506794379000 ]
```

So I don't see any issues here or anything that is likely to be an issue.  Nice work.  :star2: