@samouri 2017-11-17
a 10 line change that made a significant perf bottleneck 1000x times faster?
not bad :)